Allow gui to be started from the ibridges shell
