+++ ibridges/cli/shell.py
@@ -13,13 +13,13 @@ except ImportError:
 
 from ibridges.cli.data_operations import CliDownload, CliMakeCollection, CliRm, CliSync, CliUpload
 from ibridges.cli.meta import CliMetaAdd, CliMetaDel, CliMetaList
-from ibridges.cli.navigation import CliCd, CliList, CliPwd, CliSearch, CliTree
+from ibridges.cli.navigation import CliCd, CliGui, CliList, CliPwd, CliSearch, CliTree
 from ibridges.cli.util import cli_authenticate
 from ibridges.path import IrodsPath
 
 ALL_BUILTIN_COMMANDS=[CliList, CliPwd, CliTree, CliMetaList,
                       CliMetaAdd, CliMetaDel, CliMakeCollection, CliDownload,
-                      CliUpload, CliSearch, CliCd, CliRm, CliSync]
+                      CliUpload, CliSearch, CliCd, CliRm, CliSync, CliGui]
 IBSHELL_HISTORY_FILE = Path.home() / ".ibridges" / ".shell_history"
 
 class IBridgesShell(cmd.Cmd):
